Move money functions to lib-common
