+++ frontend/src/e2e-test/pages/employee/fridge-head-information/fridge-head-information-page.ts
@@ -7,21 +7,7 @@ import {
   selectFirstComboboxOption,
   waitUntilScrolled
 } from '../../../utils/helpers'
-
-function formatCents(amount?: number): string | undefined {
-  if (amount === undefined || amount === null) {
-    return undefined
-  }
-
-  const euros = amount >= 0 ? Math.floor(amount / 100) : Math.ceil(amount / 100)
-  const cents = Math.abs(amount % 100)
-
-  if (cents !== 0) {
-    return `${euros},${cents.toString().padStart(2, '0')}`
-  } else {
-    return euros.toString()
-  }
-}
+import { formatCents } from 'lib-common/money'
 
 interface VerifyFamilyPersonOpts {
   personId: string
