+++ frontend/src/employee-frontend/components/invoice/Sum.tsx
@@ -5,7 +5,7 @@
 import React from 'react'
 import styled from 'styled-components'
 import { useTranslation } from '../../state/i18n'
-import { formatCents } from '../../utils/money'
+import { formatCents } from 'lib-common/money'
 
 interface Props {
   title: 'rowSubTotal' | 'familyTotal'
