+++ frontend/src/employee-frontend/components/invoices/Invoices.tsx
@@ -26,7 +26,7 @@ import NameWithSsn from '../common/NameWithSsn'
 import ChildrenCell from '../common/ChildrenCell'
 import { InvoiceSummary } from '../../types/invoicing'
 import { Result } from 'lib-common/api'
-import { formatCents } from '../../utils/money'
+import { formatCents } from 'lib-common/money'
 import Tooltip from '../../components/common/Tooltip'
 import { StatusIconContainer } from '../common/StatusIconContainer'
 import colors from 'lib-customizations/common'
