+++ frontend/src/employee-frontend/components/person-profile/income/income-table/IncomeSum.tsx
@@ -5,7 +5,7 @@
 import React from 'react'
 import styled from 'styled-components'
 import { useTranslation } from '../../../../state/i18n'
-import { formatCents } from '../../../../utils/money'
+import { formatCents } from 'lib-common/money'
 
 const Container = styled.div`
   display: flex;
