+++ frontend/src/employee-frontend/components/reports/VoucherServiceProviders.tsx
@@ -28,7 +28,7 @@ import { CareArea } from '../../types/unit'
 import { getAreas } from '../../api/daycare'
 import { FilterLabel, FilterRow, TableScrollable } from './common'
 import { FlexRow } from '../common/styled/containers'
-import { formatCents } from '../../utils/money'
+import { formatCents } from 'lib-common/money'
 import { useSyncQueryParams } from '../../utils/useSyncQueryParams'
 import { defaultMargins, Gap } from 'lib-components/white-space'
 import { FixedSpaceRow } from 'lib-components/layout/flex-helpers'
