+++ frontend/src/employee-frontend/components/voucher-value-decision/VoucherValueDecisionIncomeSection.tsx
@@ -10,7 +10,7 @@ import { H3, H5 } from 'lib-components/typography'
 import { useTranslation } from '../../state/i18n'
 import { Income } from '../../types/income'
 import { VoucherValueDecisionDetailed } from '../../types/invoicing'
-import { formatCents } from '../../utils/money'
+import { formatCents } from 'lib-common/money'
 import { formatName, formatPercent } from '../../utils'
 import { Result, Success } from 'lib-common/api'
 import { useRestApi } from 'lib-common/utils/useRestApi'
