+++ frontend/src/employee-frontend/components/voucher-value-decision/VoucherValueDecisionSummary.tsx
@@ -13,7 +13,7 @@ import VoucherValueDecisionCoPaymentSection from './VoucherValueDecisionCoPaymen
 import VoucherValueDecisionValueSection from './VoucherValueDecisionValueSection'
 import { useTranslation } from '../../state/i18n'
 import { VoucherValueDecisionDetailed } from '../../types/invoicing'
-import { formatCents } from '../../utils/money'
+import { formatCents } from 'lib-common/money'
 
 type Props = {
   decision: VoucherValueDecisionDetailed
