+++ frontend/src/employee-frontend/components/voucher-value-decisions/VoucherValueDecisions.tsx
@@ -24,7 +24,7 @@ import { VoucherValueDecisionSummary } from '../../types/invoicing'
 import { SearchOrder } from '../../types'
 import { Result } from 'lib-common/api'
 import { formatDate } from 'lib-common/date'
-import { formatCents } from '../../utils/money'
+import { formatCents } from 'lib-common/money'
 import { SortByVoucherValueDecisions } from '../../api/invoicing'
 import Pagination from 'lib-components/Pagination'
