+++ frontend/src/lib-common/money.spec.ts
